requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
peft>=0.4.0
datasets>=2.12.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
python-multipart>=0.0.6
aiofiles>=23.0.0